Move argument-parser to config
This is part of a larger effort to clean up the main source file.
